Move instruction table to `Assembler.Table`
